Fixed Bugs #1002 and #971 and updated the Polish Translation plus patches from Craig and Tsoots
git-svn-id: svn://scribus.net/trunk/Scribus@415 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/gettext/csvim/csvim.cpp
@@ -160,7 +160,7 @@ void CsvIm::parseLine(const QString& line, bool isHeader)
 			data += "\t" + tmp.stripWhiteSpace();
 			++colIndex;
 		}
-		else if (end == line.length() - 1)
+		else if (end == static_cast<int>(line.length()) - 1)
 		{
 			QString tmp = line.mid(start, end - 1);
 			if (isHeader)
